instructive sidebar, style updates
